app/requirements.txt
fastapi
uvicorn
chromadb-client
langchain
langchain-community
langchain-ollama
httpx
onnxruntime
tokenizers
langchain-core
aiofiles